feat: remove Custom JwtBearerHandler and add Authorize with role from all controllers

+++ Controllers/UtilisateurController.cs
@@ -1,4 +1,6 @@
+using System.Security.Claims;
 using backend_tpgk.Dtos;
+using backend_tpgk.Requirement;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -16,6 +18,7 @@ public class UtilisateurController : ControllerBase
         _utilisateurService = utilisateurService;
     }
 
+    [Authorize(Roles = "Responsable, Admin")]
     [HttpGet]
     public async Task<ActionResult<ServiceResponse<List<Utilisateur>>>> GetAll()
     {
@@ -25,9 +28,20 @@ public class UtilisateurController : ControllerBase
     [HttpGet("{id}")]
     public async Task<ActionResult<ServiceResponse<Utilisateur>>> GetSingle(Guid id)
     {
+        bool roleIsClient = false;
+        string uuid = "";
+
+        foreach(Claim claim in User.Claims){
+            if(claim.Type == "Id") uuid = claim.Value;
+            if(claim.Type == "Role" && (claim.Value == "Client" || claim.Value == "Modérateur" || claim.Value == "Assistant")) roleIsClient = true;
+        }
+
+        if(roleIsClient && uuid != id.ToString()) return new ForbidResult();
+        
         return Ok(await _utilisateurService.GetUtilisateurById(id));
     }
 
+    [AllowAnonymous]
     [HttpPost]
     public async Task<ActionResult<ServiceResponse<Utilisateur>>> AddUtilisateur([FromBody] Utilisateur body)
     {
@@ -37,11 +51,32 @@ public class UtilisateurController : ControllerBase
     [HttpPut("{id}")]
     public async Task<ActionResult<ServiceResponse<Utilisateur>>> UpdateUtilisateur(Guid id, [FromBody] UtilisateurDtos body)
     {
+        bool roleIsClient = false;
+        string uuid = "";
+
+        foreach(Claim claim in User.Claims){
+            if(claim.Type == "Id") uuid = claim.Value;
+            if(claim.Type == "Role" && (claim.Value == "Client" || claim.Value == "Modérateur" || claim.Value == "Assistant")) roleIsClient = true;
+        }
+
+        if(roleIsClient && uuid != id.ToString()) return new ForbidResult();
+
         return Ok(await _utilisateurService.UpdateUtilisateur(id, body));
     }
 
+    [Authorize(Roles = "Admin")]
     [HttpDelete("{id}")]
     public async Task<ActionResult<ServiceResponse<Utilisateur>>> DeleteUtilisateur(Guid id){
+        bool roleIsClient = false;
+        string uuid = "";
+
+        foreach(Claim claim in User.Claims){
+            if(claim.Type == "Id") uuid = claim.Value;
+            if(claim.Type == "Role" && (claim.Value == "Client" || claim.Value == "Modérateur" || claim.Value == "Assistant")) roleIsClient = true;
+        }
+
+        if(roleIsClient && uuid != id.ToString()) return new ForbidResult();
+
         return Ok(await _utilisateurService.DeleteUtilisateur(id));
     }
